Setup track manager/editor and got arduino config sending time data in the track editor
Merged from feature branch:

commit 2bcc4b3ca50da7be87ed84dfb0bd7b82c2c7c9ca
Author: Dave Aufderheide <daufderh@hotmail.com>
Date:   Sun Feb 8 23:51:31 2026 -0500

    Added logging to arduino protocol and serial connection
    Fixed arduino protocol so that at least on the track editor it makes a connection and receives heartbeat data

commit 31e50b1770136b704afc52f03a89ce48beaf8c71
Author: Dave Aufderheide <daufderh@hotmail.com>
Date:   Sun Feb 8 13:02:31 2026 -0500

    Added normally closed relays to the arduino config.  It's not used but it's there

commit 19f9c7fa4d212648c9246b907aa37827b09fc6f3
Author: Dave Aufderheide <daufderh@hotmail.com>
Date:   Sun Feb 8 00:06:10 2026 -0500

    Added ability to toggle write pins in the arduino protocol

commit 10a8b13625fe81fb3b40611b81073b989ff21b88
Author: Dave Aufderheide <daufderh@hotmail.com>
Date:   Sat Feb 7 23:23:39 2026 -0500

    Refactored track manager/editor so the arduino summary/editor are split out
    Added new testing for the new components
    Fixed some server side unit tests that were broken by previous chnges

commit e0acf62898727f413ce8998f8ac8f0bc9e83cd49
Author: Dave Aufderheide <daufderh@hotmail.com>
Date:   Sat Feb 7 21:04:06 2026 -0500

    Refactored track manager/editor to split out the arduino components
    Fixed issue with arduino callbutton summary count

commit df740d7c9dde4f054738442a659a74bbecb078f3
Author: Dave Aufderheide <daufderh@hotmail.com>
Date:   Sat Feb 7 20:01:23 2026 -0500

    Simplified pin read/write detection by storing the behaviors in a map

commit 10e9b47ea3c20df37b9563185b9600dc23180c2a
Author: Dave Aufderheide <daufderh@hotmail.com>
Date:   Sat Feb 7 19:54:31 2026 -0500

    Added write mode relay pins to the arduino config

commit cb2b572d9a333ce23f9fcacbe215cef4cdc55312
Author: Dave Aufderheide <daufderh@hotmail.com>
Date:   Sat Feb 7 19:01:06 2026 -0500

    Added reserved arduino pin config and rearranged the selector slightly

commit f8ec91465815dd6526e7a062971e01e87e228552
Author: Dave Aufderheide <daufderh@hotmail.com>
Date:   Sat Feb 7 18:27:56 2026 -0500

    Added support to update the arduino config in the track-editor

commit d67bc2c34aac1fe82b332980c20277e69f75ffd7
Author: Dave Aufderheide <daufderh@hotmail.com>
Date:   Sat Feb 7 17:40:54 2026 -0500

    Fixed track editor so it shows available com ports and pin behaviors

commit 104290336e19e41131d7b8e6911a85b51fda53f4
Author: Dave Aufderheide <daufderh@hotmail.com>
Date:   Sat Feb 7 17:16:39 2026 -0500

    Updated restart heat state so things like defer heat will work right

commit 770b6323417cb2241741fa00c2690d4ff054aa4a
Author: Dave Aufderheide <daufderh@hotmail.com>
Date:   Sat Feb 7 16:24:21 2026 -0500

    First pass at state management on the race day screen

commit d7f91be69c6e06e12f1efc5cecf7c3e76c16a573
Author: Dave Aufderheide <daufderh@hotmail.com>
Date:   Sat Feb 7 00:51:13 2026 -0500

    Added ack screen
    Race day now monitors interface status

commit a3d8016aab3f04a70d902f29714e0a243b0d1291
Author: Dave Aufderheide <daufderh@hotmail.com>
Date:   Fri Feb 6 17:19:58 2026 -0500

    Added heartbeat from the track interface to the client

commit 16caaed0e41e2664c169f3693a133c0fc7989e19
Author: Dave Aufderheide <daufderh@hotmail.com>
Date:   Thu Feb 5 23:45:42 2026 -0500

    Fixed com port selector when first entering track editor

commit beade924eba068519bd4350216e04c528acb9ea5
Author: Dave Aufderheide <daufderh@hotmail.com>
Date:   Thu Feb 5 22:35:23 2026 -0500

    Client track editor trying (and failing) to initialize an interface when entered

+++ server/src/main/java/com/antigravity/service/DatabaseService.java
@@ -93,6 +93,10 @@ public class DatabaseService {
 
         private Track resetTracks(MongoDatabase database) {
                 MongoCollection<Track> trackCollection = database.getCollection("tracks", Track.class);
+                if (trackCollection.countDocuments() > 0) {
+                        return trackCollection.find().first();
+                }
+
                 trackCollection.drop(); // Clear all existing data
 
                 // Reset sequence
@@ -113,7 +117,9 @@ public class DatabaseService {
                 Lane l4 = new Lane("#fbbf24", "black", 100, getNextSequence(database, "lanes"), null);
                 lanes.add(l4);
 
-                Track track = new Track("Bright Plume Raceway", lanes, getNextSequence(database, "tracks"), null);
+                com.antigravity.protocols.arduino.ArduinoConfig config = new com.antigravity.protocols.arduino.ArduinoConfig();
+                Track track = new Track("Bright Plume Raceway", lanes, config, getNextSequence(database, "tracks"),
+                                null);
 
                 trackCollection.insertOne(track);
                 System.out.println("Tracks reset.");
